App Extension support
This adds workarounds for the code that was preventing React from compiling when linked against an iOS App Extension target.

Some iOS APIs are unavailable to App Extensions, and Xcode's static analysis will catch attempts to use methods that have been
flagged as unavailable.

React currently uses two APIs that are off limits to extensions: `[UIApplication sharedApplication]` and `UIAlertView`.

This commit adds a helper function to `RCTUtils.[hm]` called `RCTRunningInAppExtension()`, which returns `YES` if, at runtime,
it can be determined that we're running in an app extension (by checking whether the path to `[NSBundle mainBundle]` has the
`"appex"` path extension).

It also adds a `RCTSharedApplication()` function, which will return `nil` if running in an App Extension.  If running in an App,
`RCTSharedApplication()` calls `sharedApplication` by calling `performSelector:` on the `UIApplication` class.  This passes
the static analysis check, and, in my opinion, obeys the "spirit of the law" by only calling `sharedApplication` after we've
determined that it's kosher to do so.

Existing calls to `[UIApplication sharedApplication]` have been replaced with calls to `RCTSharedApplication()`, with `nil` checks
where appropriate.

The absence of `UIAlertView` is similarly handled with a runtime workaround.  If we're running in an App, the
 `RCTAlertView(NSString *title, NSString *message, id delegate, NSString *cancelButtonTitle, NSArray *otherButtonTitles)`
function uses `NSInvocation` to call the extension-unavailable `initWithTitle:message:delegate:cancelButtonTitle:otherButtonTitles:`
selector.  In an Extension it just logs an error and returns `nil`.

As a result of these changes, React can be linked into a shared library that is then linked into both an App and an App Extension.
When running in the App, all existing functionality will continue to work as before.  When running in the Extension, `AlertIOS` will
be unavailable, as will APIs that depend on access to `[UIApplication sharedApplication]`.

APIs that are unavailable to extensions are:
`RCTActionSheetManager` - uses `sharedApplication` to access the `window` of the app delegate
`RCTImagePickerManager` - uses `sharedApplication` to get the `keyWindow` property
`RCTLinkingManager` - uses `sharedApplication` to open URLs
`RCTPushNotificationManager` - uses `sharedApplication` to badge the app icon and register for permissions
`RCTDevMenu` - uses `sharedApplication` to get the `keyWindow` property
`RCTPerfStats` - uses `sharedApplication` to access the `window` of the app delegate